CT-1776 return json and use js promises
At first we were rendering js files in the action controller and passing
the content through these rendered methods. Ideally all we care about is
the content that the controllers action will return. Therefore a better
format would be to return JSON and let the frontend handle it.

this commit also introduces javascript promises to handle scenarios where
validations fail,