CT-1768 Redesign Man stats page
This uses a new field on ReportType to indicate whether or not
it is a standard report; this, and the foi, sar columns govern where
the report appears on the page.

Furthermore, each report types now has a default reporting period column,
which can be year_to_date, quarter_to_date, etc.  This is used to
indeicate the reporting period on the main page, as well as the reporting
period if the report is called without period_start and period_end parameters.