CT-1869 Fix bug caused by deleting linked cases
If a case had linked cases, and one of the linked cases was subsequently
soft deleted, that case became invalid, and could no longer be updated due
to failing validation errors.

This is fixed by the removal of the  `vaidates :linked_case` on the 'LinkedCase'
model.  This is not required, as there are validations for either linked_case id or
linked case number on creation.