CT-1808 Bugfix for users with multiple roles
A bug was reported where users with multiple roles (manager, responder and approver)
were unable to mark a case as responded.

This is caused by the Case::Base#team_for_user method which was used to determine
which team to pass into the state machine as the acting team.  It just took the
first team that was common to both the case (i.e. on the assignements) and the User#team_roles.
This took no account of the role of the team, therefore we had the clash of StateMachine
events being called with a team which had no rights to perform that event.

This has been replaced by two methods:

 - Case::Base#team_for_assigned_user(user, role)
 - Case::Base#team_for_unassigned_user(user, role)

Note that both methods now take role as a parameter, to ensure that the correct matching team
is returned.

the difference between the two methods are that #team_for_assigned_user looks for an intersect between
the user's teams, and the accepted assignments for that user on the case, whereas #team_for_unassigned_user
looks for an intersect between the user's teams and teams on assignments in  unassigned or pending (not necessarily
for the same user).