Fix Dev seeder
The rake task `db:reseed` was not working correctly - the business
units were not able to process Overturned SAR correspondence types.

This was because the data migration PopulateTeamsForOverturnedSars
was being run before the teams had been populated, and therefore had
no effect.  This was fixed by adding an additional data migration to
run the DevUserSeeder instead of running it in the rake task itself,
with a check to ensure it doesn't run on production.