CT-1876 allow managers to extend for pit instead of specialist
An issue arose where non_trigger FOIs could not be
extended_for_pit since they needed an assigned_disclosure_specialist
and as they have no disclosure approvers the action was not permitted.
To solve this problem we have changed the extend for pit action to be
permitted to any manager on a case.