CT-1919 Case::Base subtypes specify which state machine config to use
Previously, each Case::Base sub-class used a state machine config based on the type abbreviation.
That meant a separate state machine config for each subclass.  This PR implements a #state_machine
method on Case::Base which is used to determine which state machine config to load.  This is set to use the
type bbreviation, but is overridden in Case::OverturnedICO::SAR to use the sar state machine config.