CT-1944 Fix bug where OVT SAR cases appeared to link to themselves
This PR fixes a bug caused by passing in the original case to the
linked case partial (rather than the current case).  This then displayed
the current case as a linked case.

In fixing this bug, we also found that the original ICO appeal was not being
displayed in Other Related cases section.  This is also fixed by creating a
new scope to include the original ico appeal as well as related cases.