CT-1937 Fix Case Linking bug
This bug was caused by a user being both a responder and a manager.
Only a manager has permissions to link SAR cases (and OVT SAR cases),
but in certain circumstances, the 'User#.teams_for_case' returns a
responding team first, which causes the error.  This only seems to happen
in production environments where the database can return records in a
different order than on the local database.

The fix is to ensure only manging teams are selected as teh acting team
in the CaseLinkingService for SAR and OVT SAR cases.