CT-1965 Hide request further clearance message
When we go live with non-trigger OVT FOI, there will be a period
in which trigger OVT FOI will not be available, so we should suppress
the Request further clearance link.

This is done by testing the overturned_trigger_fois FeatureSet,
and only enabling that action in the policy if the feature set
is enabled.

Once triggeer OVT FOIs have been rolled out, the request_furnter_clearance?
method in polocies/case/overturned_ico/foi_policy.rb can be removed, and let the
superclass handle the policy.