CT-2009 fixes issues with dropzone not displaying filename
Users reported that they were not able to upload files. What was happening
is that dropzone was not displaying the filename and file size for the file
that was uploaded. The cause of this issue was from us recently updating
slim-rails and slim changed the way html data attributes were output. This
meant that dropzoneJS could not find the right attributes to update and
display.

https://github.com/slim-template/slim/issues/807
https://github.com/slim-template/slim/issues/820 someone else reported
this issue to slim and suggested adding a configuration option to allow
underscore to hyphen conversion to still take place.

This commit updates dropzoneJS to the latest version.