CT-2025 Background creation of report R005
Currently, there is no configuration for R005 in the config/sidekq-backgound-jobs.yml
which means when the stats controller gets a request to donwload R005, it will always
generate a new report, rather than look to see if there is anb existing one in the database.

This report is now taking too long to generate and the request times out.

This adds a configuration to generate the rport every hour (at five past the hour),
and so shuold obviate this problem