CT-2077 Add temporary limit to closed case CSV output
There are around 8,000 closed cases on the live system, and
requesting a CSV download of all thses was causing server timeout
on live.  While investigation is continuing on how to stream these
in a production environment, a temporary fix to limit output to
the first 1000 cases (by last transitions date in descending order)
has been implemented.