CT-2060 Upload durring escalation period bugfixe
REsponders should not be permitted to upload a response until 3
days (the escalation perios) has elapsed to give approvers enough
time to view the dase and take it on if they deem necessary.

There was a bug which allowed them to upload the response during
that escalation period.  This PR fixes that bug.