CT-2280 offender sars validation (#822)
Performing the correct minimal validation on each screen of the create case flow for Offender SAR

This is the reason why we have the form object; it does the work of displaying each step of the wizard and knows which validations are required on each step. We have all the validations in the Case::SAR::Offender model, but we have monkey patched ApplicationRecord with a valid_attributes? method which takes a hash of params to validate using the model's validators as opposed to calling valid? which validates all the attributes in one go. By definition we need to only validate certain fields at each step.

Because of the way radio button groups work, nothing is sent in the params if neither button is selected, which defeats the partial validation which wants to validate only fields that are passed on each form submission. Accordingly, and rather annoyingly, we need to merge in some empty params if the radio group is empty.

Also, because of way we partial validation using the fields submitted, we need to unset values for fields from the params whose validation is conditional on the values of other fields. 

https://dsdmoj.atlassian.net/browse/CT-2280