CT-2238 new offender sar workflow states (#852)
Implement support for the available states for Offender SARs and provide an interface to move it through the workflow states. A newly created case should be able to be viewed and start off in "Data to be requested". You can then use the green button to move it through the states. 

* CT-2238 Add basic states for offender sars workflow

* CT-2238 Added cases controller with create action

* CT-2238 Amended logic to use same controller action for ratcheting state

* CT-2238 Add additional fields to the search list

* CT-2238 Case transitions working up to close case

* CT-2238 Remove ready to close and add ready to copy

* CT-2238 Dry up the controller spec using a shared example

* CT-2238 Remove stray binding.pry statement and add todo

* CT-2238 Add support for closing the case

* CT-2238 Refactor state transitions into offender sar controller