CT-2257 add note to offender sar (#853)
CT-2257 add note to offender sar 

Allow Branston users to add notes to cases that get displayed inline in the case history. Notes are stored using the message property in the JSON field for the case, and get displayed as the details element when case transition items render in the case history. Normal conversation messages are still filtered out based on their different event name, so if we were to add the conversation feature back for offender SAR, it should still work.

* CT-2257 Add note to case

* CT-2257 Add validation to blank message field in case browser doesn't support required

* CT-2257 move govuk date fix into its own concern

* CT-2257 Refactor notes and messages controllers to use notable concern