Ct 2423 generate acknowledgement letter text (#922)
Letter templates and letter generation based on a new database model for letter templates. Templates use ERB for replaceable fields since it's built in to Ruby hence no new dependencies in the app.

The letter_template takes a values object, which can either be a case or specially constructed - we'll be adding a Letter class which wraps a case in order to add features subsequently (such as knowing when a previous letter was sent and referring to it in a template), but for now we're passing in the case directly. The letter is just used presently to drive Rails' `form_for`.

Letter templates have different types; this includes the first 'acknowledgement' letter which can be generated either directly after creation or while waiting for data.