CT-2228 Basic word export working (#956)
User needs to be able to download the generated letter as a Word document. Added an alternate rendering format for :docx, a link hitting same, and amended the show action to produce the document. Some basic tests and a little refactoring of the Letter class so it now wraps letter template and Case.