CT-2604 Deployment, replication and rolling deployment (#982)
Updating deployment scripts - this PR provides two shell scripts, one to build, tag and push an image of the desired branch up to the ECR, and the other to take the tag for that image and deploy it to a given namespace. 

This PR also adds more replicas and asks for rolling restarts. 

Build and deploy usage looks like this:

    # build the app to get an image tag
    ./build.sh
    ...many lines of output...
    Image created with tag: cts-CT-1234-my-lovely-branch-6bece953

    # deploy image-tag to dev
    ./deploy.sh cts-CT-1234-my-lovely-branch-6bece953 dev

    # deploy latest image of master to production
    ./deploy.sh cts-master-8f3tgba3 production