Make new team's name equal the original team
The Problem:
Active team's names must be unique, yet the new team has to have the same name as the original,
 and we need to save the new team before deactivating the original
 (deactivating the original team will change it's name to include 'Deactivated' text)

The Solution:
Change the new team's name initially such that the team can be saved without violating the 'unique' validation.
After deactivation of the original team, which changes it's name, set the new team name equal to the original name of the original team.