CT-2223 Edit case details (#1068)
This introduces the ability to edit Offender SAR cases, adding "change" links for each group of fields corresponding to the grouping on the Create Case journey

The form partials are common to both create and edit journeys, and the back link is now a helper so that it can be told which context it's to work in. In the create context, it works out the previous step from the "steps" defined for the form object.