CT-2928 add new fields for contact details (#1183)
* CT-2928 added new methods for accessing contact details

* CT-2928 Recipient added

* CT-2928 external reference - retaining the old json field because otherwise in-progress sessions break

This PR includes the addition of the recipient enum field, renames the third_party_reference to requester_reference and adds convenience methods to make it easier to get recipient and requester details for subsequent use.

The values for the enum are appended _recipient because the model gets automatically created convenience helper methods including a predicate method. When we originally proposed the values, these included "third_party" which generated a convenience helper method third_party? which overwrote an existing predicate method and let to some confusing behaviour. So we've namespaced the values as described to prevent this collision. Now the new predicate method from the enum is kase.third_party_recipient?