CT-2919 update letter templates (#1195)
* CT-2919 Add letter_recipient and letter_address methods

* CT-2919 Handle third party without name properly

This PR adds support for letter_recipient which is a convenience method on the Letter object which calls either "requester_name" or "recipient_name" depending on whether it is an acknowledgement letter or a dispatch letter, and the same for letter_address. This allows us to keep logic out of the letter templates themselves and have simple method calls in the ERB templates for those.