CT-3215 Ensure assignments are present before trying to build link that relies on them (#1364)
After moving people/business units around in the CMS and deactivating the "Private Office" team this resulted in some users being unable to view a large number of cases.

On investigation, one problem seemed to be the deactivation of the private office team, which has a specific value for "Code" that had been amended as part of the deactivation.

We restored the missing team, making no other changes to the database. This restored the ability to access some of the cases but other cases failed with a different problem now.

[09120983cf00da9c7d7e94c6c5c62bff] ActionView::Template::Error (No route matches {:action=>"reassign_user", :case_id=>"20390", :controller=>"assignments", :id=>nil}, missing required keys: [:id]):

This happens in the cases helper for reassign user when the user has no assignments. The helper assumes there will always be at least one assignment or more, but in this case on production it seems there are none and the helper requires one. This PR adds a guard clause to ensure there is always at least one and return an empty string if not.