Bug fix and add test to cover it
Fix bug so when you change the contact details but not the type the
status doesn't reset.

Was due to the difference in comparing @kase.complaint_type (say 'ICO')
and @params['complaint_type'] which returns 'ico_complaint'. Added a
mapping function to get around this distinction and a test to the
service class to cover it.