Add extra step to feature test
Covers the case where the contact details for a case type are change but
not the type itself. This test will fail if the wording of the
underlying names of the types change so it was a good addition to the
unit test the ComplaintCaseUpdaterService spec that also covers this
scenario.