CT-3448 fix migration bugs (#1499)
* Specify Rails version on migrations to avoid error

Error message was:

Directly inheriting from ActiveRecord::Migration is not supported. Please specify the Rails release the migration was written for

* Fix migration error where class was being passed

a string of the class name was expected instead.

This was the error solved by this commit:

ArgumentError: A class was passed to `:class_name` but we are expecting
a string.