Add less cmd tool into build (#1549)
Adds less into Dockerfile. Rails console "rails c" depends on pry which
in turn depends on less. Without this it means we can't use a lot of
rails console stuff in environments.

For instance Model.where(something: "thing") does not work.

Adding back this into the build to try to fix this issue.