Add "your reference" to word letter template
This is a change to the actual word docx template rather than the html
body content which is defined in the letter seeders.

To get this to work you have to add a new 'merge field' through the word
GUI and not just copy an existing reference in the doc as this won't be
picked up.

In the doc there are references like «=letter_address» which are merge
fields, if you copy paste this and just change the name it won't work.

You need to go to on word for Mac: insert > field, then select 'Mail
Merge' > 'MergeField' then in the input on the dialog that has
'MERGEFIELD' add '=name_of_your_new_merge_feild', so it looks like:

'MERGEFIELD =name_of_your_new_merge_feild'

then > click 'OK'.

Then save the doc.

You then pass the new param to the template_data hash passed to the
Sablon template in LettersController and it should render on the word doc download.

If you don't follow these steps it won't work...