Ct 3693 - xss brakeman (#1638)
* Avoid xss when using raw to product letters

letters are seeded to the DB as strings that are interpreted at view
compile time and dynamic values are added. use of raw is needed here as
it means the slim / ERB engine will treat the string as a view rather
than a literal string.

The issue with raw is that it doesn't escape the values, meaning that
xss is possible. This commit clones (this is sothe original case isn't updated)
the original case object and escapes the values used in the letter so that
this isn't possible.

* Update brakeman ignore