Add brakeman Dangerous Send to ignore list.
This is safe to ignore as there is a check on the
events/methods that are permitted, which means that bad
data can't be sent to the .send() on the state machine.

this is unsafe in the docs:
method = params[:method]
@result = User.send(method.to_sym)

and you would resolve it as:
method = params[:method] == 1 ? :method_a : :method_b
@result = User.send(method, *args)

but in our code we effectively have:

if the_param_with_method_name_is_in_the_permitted_list
  @obj.send(params[:permitted_method_name])

So if someone tried to pass :delete_all it would not reach the @obj
, as it is not in the allowed list.

In our code there is also a catch which raises an argument error
if the method is not permitted.