Add ignore to brakeman Redirect warning
The fear with this warning is that user input is used to
create a redirect URL that is melicious.

However in the case here it is not user input that is used to
create the url. Rather it is a URL that is generated internally
by the CaseAttachments model to point to the the actual attachments
URL in Amazon S3 to allow it to be downloaded.

Given this we can safely ignore this warning in this instance.