Add Command Injection warning to ignore file
The potential issue here is that a melicious use could
pass dengerous arguments appended to the system call.

However in this instance no user input is supplied ot the
system() method and all arguments are either constants
designated in code or generated without reference to outside the class.

Similarly the commands to actually run are defined within
the method itself so are not passed in or manipulatable
outside of the method.

Of the two arguments used in the system call:

"folder" - this is a folder name generated in a private
           method using a SecureRandom value in the same
           class as the system call.

"self.result_name" - this is defined in child any child
                     class as a constant that has no user
                     input.

Therefore the system call is safe.