Add ignores for Dynamic render paths
The issue here is that a param is passed that could be used to get
to a page the user shouldn't if they pass a different param than
what they are supposed to.

In our case though the acceptable params are checked against a list of
acceptable parameters. This means that if someone was to pass an action
that is not defined here an error would be thrown - as such it is a
false positivek