Add ignores for two dynamic render path brakeman warnings
For app/views/admin/cases/new.html.slim:15

Values here come from the case type abreviation which
map to specific partials.

If non matching strings are passed to the path to render
the partial from a query param then an error is thrown.

Therefore this is a false positive.

-----

For app/views/cases/show.html.slim:72

This page also gets the value it uses to render the partial
the case_details partial from the case type_abreviation

However this is rendered as part of the interpretation of a
closing view and is as such not a parameter that a person
could pass to access partials that they would not usually
have access to.

Hence this is a false positive as well.