Ct 3698 breakman fixes (#1745)
* Add brakeman Dangerous Send to ignore list.

This is safe to ignore as there is a check on the
events/methods that are permitted, which means that bad
data can't be sent to the .send() on the state machine.

this is unsafe in the docs:
method = params[:method]
@result = User.send(method.to_sym)

and you would resolve it as:
method = params[:method] == 1 ? :method_a : :method_b
@result = User.send(method, *args)

but in our code we effectively have:

if the_param_with_method_name_is_in_the_permitted_list
  @obj.send(params[:permitted_method_name])

So if someone tried to pass :delete_all it would not reach the @obj
, as it is not in the allowed list.

In our code there is also a catch which raises an argument error
if the method is not permitted.

* Add ignore to brakeman Redirect warning

The fear with this warning is that user input is used to
create a redirect URL that is melicious.

However in the case here it is not user input that is used to
create the url. Rather it is a URL that is generated internally
by the CaseAttachments model to point to the the actual attachments
URL in Amazon S3 to allow it to be downloaded.

Given this we can safely ignore this warning in this instance.

* Add Command Injection warning to ignore file

The potential issue here is that a melicious use could
pass dengerous arguments appended to the system call.

However in this instance no user input is supplied ot the
system() method and all arguments are either constants
designated in code or generated without reference to outside the class.

Similarly the commands to actually run are defined within
the method itself so are not passed in or manipulatable
outside of the method.

Of the two arguments used in the system call:

"folder" - this is a folder name generated in a private
           method using a SecureRandom value in the same
           class as the system call.

"self.result_name" - this is defined in child any child
                     class as a constant that has no user
                     input.

Therefore the system call is safe.