Add remote code execution to ignore list
Code: ReportType.find_by_abbr(specific_report).class_name.constantize
File: app/controllers/application_controller.rb

Reason:
This is the name of a report which is passed in as a param
to look up a report type from the DB - this report type is
passed as a string to .find_by_attr a Rails method meaning
that melicious code would be filtered out.

If a report doesn't exist then the ReportType class will
just return nil. False postitive.