Add ignore for false positive RCE warning
file: app/views/teams/index.html.slim
line: 57

@reports is instantiated via a call to
ReportType.where(full_name: 'Business unit map),
which is hard coded in the TeamsController#index
action. This brings back a list of report records of
which the class_name field is interpreted to lookup
the correct class and a hard coded description is rendered
from the resulting class.

The DB report types are seeded on deploy and there can't
be updated from the UI. Call to .description after calling
the class name means a class has to exist in the application
for this to function.

False positive.