CT-3696 brakeman fixes (#1758)
* Add remote code execution to ignore list

Code: ReportType.find_by_abbr(specific_report).class_name.constantize
File: app/controllers/application_controller.rb

Reason:
This is the name of a report which is passed in as a param
to look up a report type from the DB - this report type is
passed as a string to .find_by_attr a Rails method meaning
that melicious code would be filtered out.

If a report doesn't exist then the ReportType class will
just return nil. False postitive.

* Add ignore for false positive RCE warning

file: app/views/teams/index.html.slim
line: 57

@reports is instantiated via a call to
ReportType.where(full_name: 'Business unit map),
which is hard coded in the TeamsController#index
action. This brings back a list of report records of
which the class_name field is interpreted to lookup
the correct class and a hard coded description is rendered
from the resulting class.

The DB report types are seeded on deploy and there can't
be updated from the UI. Call to .description after calling
the class name means a class has to exist in the application
for this to function.

False positive.

* Add ignore for false positive RCE warning

file: app/views/teams/show.html.slim
line: 34

@reports is instantiated via a call to
ReportType.where(full_name: 'Business unit map),
which is hard coded in the TeamsController#show
action. This brings back a list of report records of
which the class_name field is interpreted to lookup
the correct class and a hard coded description is rendered
from the resulting class.

The DB report types are seeded on deploy and there can't
be updated from the UI. Call to .description after calling
the class name means a class has to exist in the application
for this to function.

False positive.