Remove call to validate steps if case is final step
This avoids having to add a lot of validation functions in the parent
'Standard' class as step validations are really simple (just link /
confirm step now) and not contingent on conditional logic. Conditional
logic was not being called by .attributes_valid? in ApplicationRecord
which meant having to declare loads of validation methods in Standard
SAR class.

This rework gets around this by not calling the step validations on the
final step and instead just validating the case completely.