Ct 3922 create sar ir case (#1806)
* Add method to .humanise subject type

* Add test for page rendering with add case header

* Add small case details field at top of page + test

* Basic form with partial

* Validations working as expected

* complete basic create spec

* Test for created case

* Add sar_ir_subtypes

* Display Original case details on SAR IR subject

* Make carried over attributes more testable

* Change dropzone copy

* Add default field to make SAR IRs triggered cases

Also fixes test failure issue

* Make test more resiliant

* Remove superflous validation declarations

* Fix error messages

* Fix spec

* Remove call to validate steps if case is final step

This avoids having to add a lot of validation functions in the parent
'Standard' class as step validations are really simple (just link /
confirm step now) and not contingent on conditional logic. Conditional
logic was not being called by .attributes_valid? in ApplicationRecord
which meant having to declare loads of validation methods in Standard
SAR class.

This rework gets around this by not calling the step validations on the
final step and instead just validating the case completely.

* Add text to en.yml or rely on defaults from InternalReview case

* Use hidden inputs for default fields

* Add individual page titles for steps

* Fix spec