CT-4073 Implement SSO with Azure Active Directory
This PR implements SSO user journey with Azure Active Directory, through the already existing Devise gem and its OmniAuth capability.

The feature can be enabled/disabled through the `enabled_features` mechanism configured in `config/settings.yml`
At the time of this PR it is enabled on Local, for developers to test and play with it.

For more details in Confluence:
https://dsdmoj.atlassian.net/wiki/spaces/STSDD/pages/3925574013/Single+Sign-on