Ct 4128 implement off sars closure logic (#1971)
* Basic service to calculate planned ersure dates and specs

* Add retention schedule for case closure

* Add functionality for adding retention_schedules

to offender sar complaints + linked cases

* Refactor service

* Updates related complaints for complaint cases on closure

Adds code to update other complaints created against the
Same original case as the current complaint is closed.

* Add logic so only closed related complaints

recieve retention_schedules on closure of related cases

Open complaints that are related will not recieve a retention
schedule on the closure of the related cases.

* Refactor spec

* Add some error handling

* Use factories rather than helpers

* Add opt out for n+1 issue

* Add feature spec for adding retention_schedules

This is for offender sar complaints

* Add guard clause for open cases

This will stop the services mutating open cases if they are
passed to it. This is because open cases should not have
RetentionSchedules at all.