Bump Rails and dependencies
Bump Rails to the latest patch in the 6.1.x branch, along with several other dependencies to their minor bugfix versions.

This resolves a bunch of dependabot alerts.

Updated FactoryBot custom strategies to work with latest version, as they introduced a breaking change in commit:
https://github.com/thoughtbot/factory_bot/commit/1b81d5dc258e5f1ad25c5109cb4bf3f6be83deb4

Note: as part of this work I've noticed many rendering deprecation warnings. As these are not related to this PR, I will raise a separate PR to fix them (as many as I can).