Disable dependabot version updates
Update `dependabot.yml` to only raise PR for security versions, and not for version updates, as per documentation here:

https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#open-pull-requests-limit

This is intended to reduce the amount of noisy PRs we get, so we can bump the gem versions (if convenient) at our own pace.

Also as part of this, there were already a few gems ready to bump so I did.