CT-4132 Add filter for retention state to the UI (#2005)
This PR adds a new filter to the RRD pending -> Pending removal tab, to filter by retention state, showing as options `Not set`, `Retain` and `Review` as these are the only states that will be applicable in this tab.

In another PR I will develop the other filter defined in this ticket. Didn't want to add all to this PR as it is quite big already.

I also figured out how to fix the capitalisation of RRD in the menu/tabs. And fixed a few deprecation warnings for `errors.add` as I saw them when running the tests.