Ct 4208 anonymise cases (#2009)
Adds a service that can anonymise Offender SAR and Offender SAR Complaint cases.

It also anonymises peripheral relations to a case namely:

- case history notes
- data request notes
- and removes papertrail versions

The service also updates the cases retention schedule to the terminal state of :anonymised. 

There was an issue with the names of the states 'destroyed' was conflicting with rails methods so the record was unable to save. Because of this RetentionSchedule states now use 'anonymise' as the base word though what the
user sees is the same.