CT-2206 Enable Style/MutableConstant rubocop rule (#2013)
This was an old ticket, so we either do it or remove the ticket and forget about it.
I think it is not a bad idea to do it. There are a lot of modified files but the changes are very specific.

This includes converting a mutating constant `TYPED_FILTER_FIELDS` into a class variable which is ugly but better at least keeps the functionality intact as the constant could not be frozen after applying this rubocop rule.