Update retention information in case details page
Preparation for ticket CT-4134 (Edit RRD Case Details).

Show in the case details page the retention information. I'm not adding the "Indicative SAR count" row because I believe this is the fuzzy search that has been descoped.

Also, the "Change" link is a placeholder, this will be implemented in another PR (as well as making sure it shows/hide depending on current user permissions).