Add scaffolding code to deal with form objects
This includes a superclass `BaseFormObject` and a controller concern `FormObjectUpdatable`.

Form objects will inherit from the `BaseFormObject` superclass which will make them very similar to ActiveRecord objects except without a DB table behind.

These form objects just need to conform to a simple interface, i.e. implementing a private `#persist!` method with the logic to save the attributes.

Then you call `#save` on the form object, which will run the validations and if pass, persist the attributes to the `record`, otherwise returns false.

The controller concern just simplifies common tasks like the whitelisting of form params (particularly tedious for GovUkDateFields) and also exposes a helper method `update_and_advance` to use in the controller `#update` action, passing the form object class.

If the `#update_and_advance` fails, it renders the `edit` action with the errors, otherwise it calls the block that usually will perform a redirect, with some flash message.