Tidying up some comments and methods
Also removed the `#attributes_map` instance method as it is not used in this incarnation of the code. Might be reintroduced later on if required tho.

Admittedly, much of this code is needed just to play nice with `acts_as_gov_uk_date`, as the gem was never coded to be other than with ActiveRecord models, and here we have ActiveModel objects.