CT-4134 Branston GDPR - Edit RRD state (#2023)
* Tidying up some comments and methods

Also removed the `#attributes_map` instance method as it is not used in this incarnation of the code. Might be reintroduced later on if required tho.

Admittedly, much of this code is needed just to play nice with `acts_as_gov_uk_date`, as the gem was never coded to be other than with ActiveRecord models, and here we have ActiveModel objects.

* Adds functionality to change the retention state/status

There is some validation involved to not allow to revert a schedule back to "not_set", as per acceptance criteria.

* Further streamlining of the BaseFormObject

Get rid of some code, as ActiveModel provides much of this out of the box.