CT-4134 Branston GDPR - Edit RRD permissions
Follow-up to PRs #2021 and #2023.

Implements the permissions mechanism to ensure only Branston team admins can manipulate the retention details.

Note: I presume the bulk action should also have some kind of check but I've left that out for now.

Also in this PR after a successful retention update, a note will be added to the case message history with what was changed (previous value and new values). A further refinement might be if nothing has changed, do not add a history message but that is a minor thing that can wait.