Refactor the retention schedule case notes (#2029)
As part of some previous PRs, we now have 2 places where we add retention schedule notes to the case history, and these messages relate to the same so we can unify and make the code nicer with a little presenter.

This presenter also handles the logic for which message (if any) to add to the history, as it can be only the `state` was changed but not the `planned_destruction_date` or the opposite, or even no changes at all.

This is done our of the box with ActiveRecord Dirty (`model.saved_changes` returns a map of the changes if any).

The presenter will also localise the date and translate the state. All quite easy to tweak if we want any different message in the future or even more attributes.

Note: this same mechanism can be applied to similar use-cases in the future.