CT-4222 Add 'Destroy' to GDPR Status Filter
On the ‘Pending Removal tab’ within the Filter by GDPR RRD status add the status of ‘Destroy’, so that the user can also filter for cases that are marked as Destroy on the Pending Removal tab.

This tab can contain cases in this state so it makes sense to allow this filter.