CT-4219 After retention destroy case is read only
Once a case has become anonymised, the case should be made read-only to prevent users from making any changes and adding any Personal Identifiable Information in the future.

To know if a case has been anonymised we check their retention schedule.

In addition, although there will be no UI for this, added an extra validation to the `RetentionScheduleForm` so it is not possible to change the `state` once it has reached the `anonymised` state.

This is a PR for the backend changes. In a follow-up PR the UI changes will be implemented (mainly, hiding the change links, etc.)